requirement.txt
fastapi
uvicorn[standard]
transformers
torch
sentence-transformers
faiss-cpu
gitpython
